requirements.txt
ansi2html==1.8.0
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
execnet==1.9.0
Faker==15.0.0
gmssl==3.2.1
idna==3.4
iniconfig==1.1.1
jsonpath==0.82
loguru==0.6.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycryptodome==3.15.0
pyparsing==3.0.9
pytest==7.1.3
pytest-forked==1.4.0
pytest-html==3.1.1
pytest-metadata==2.0.2
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.12